Split Engram ops into standalone files
