+++ onnxruntime/contrib_ops/cpu/bert/ngram_hash_mapping.h
@@ -6,22 +6,8 @@
 #include "core/common/common.h"
 #include "core/framework/op_kernel.h"
 
-#include <string>
-
 namespace onnxruntime {
 namespace contrib {
-
-template <typename T>
-class ShortConv final : public OpKernel {
- public:
-  explicit ShortConv(const OpKernelInfo& info);
-  Status Compute(OpKernelContext* context) const override;
-
- private:
-  std::string activation_;
-  int64_t dilation_;
-  float epsilon_;
-};
 
 template <typename T>
 class NgramHashMapping final : public OpKernel {
@@ -33,16 +19,6 @@ class NgramHashMapping final : public OpKernel {
   int64_t max_ngram_size_;
   int64_t n_head_per_ngram_;
   T pad_id_;
-};
-
-template <typename T>
-class EngramGate final : public OpKernel {
- public:
-  explicit EngramGate(const OpKernelInfo& info);
-  Status Compute(OpKernelContext* context) const override;
-
- private:
-  float epsilon_;
 };
 
 }  // namespace contrib
