+++ onnxruntime/contrib_ops/cpu/bert/short_conv.h
@@ -1,0 +1,27 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "core/common/common.h"
+#include "core/framework/op_kernel.h"
+
+#include <string>
+
+namespace onnxruntime {
+namespace contrib {
+
+template <typename T>
+class ShortConv final : public OpKernel {
+ public:
+  explicit ShortConv(const OpKernelInfo& info);
+  Status Compute(OpKernelContext* context) const override;
+
+ private:
+  std::string activation_;
+  int64_t dilation_;
+  float epsilon_;
+};
+
+}  // namespace contrib
+}  // namespace onnxruntime
