+++ onnxruntime/contrib_ops/cuda/bert/engram_gate_impl.h
@@ -11,36 +11,6 @@ namespace contrib {
 namespace cuda {
 
 template <typename T>
-Status LaunchShortConvKernel(
-    cudaStream_t stream,
-    const T* input,
-    const T* weight,
-    const T* norm_scale,
-    const T* bias,
-    T* output,
-    int64_t batch_size,
-    int64_t sequence_length,
-    int64_t hc_mult,
-    int64_t hidden_size,
-    int64_t kernel_size,
-    int64_t dilation,
-    float epsilon,
-    bool apply_silu);
-
-template <typename T>
-Status LaunchNgramHashMappingKernel(
-    cudaStream_t stream,
-    const T* input_ids,
-    const T* multipliers,
-    const T* vocab_sizes,
-    T* output,
-    int64_t batch_size,
-    int64_t sequence_length,
-    int64_t max_ngram_size,
-    int64_t n_head_per_ngram,
-    T pad_id);
-
-template <typename T>
 Status LaunchEngramGateKernel(
     cudaStream_t stream,
     const T* embeddings,
