+++ onnxruntime/contrib_ops/cuda/bert/ngram_hash_mapping.h
@@ -6,23 +6,9 @@
 #include "core/common/common.h"
 #include "core/providers/cuda/cuda_kernel.h"
 
-#include <string>
-
 namespace onnxruntime {
 namespace contrib {
 namespace cuda {
-
-template <typename T>
-class ShortConv final : public onnxruntime::cuda::CudaKernel {
- public:
-  explicit ShortConv(const OpKernelInfo& info);
-  Status ComputeInternal(OpKernelContext* context) const override;
-
- private:
-  std::string activation_;
-  int64_t dilation_;
-  float epsilon_;
-};
 
 template <typename T>
 class NgramHashMapping final : public onnxruntime::cuda::CudaKernel {
@@ -34,16 +20,6 @@ class NgramHashMapping final : public onnxruntime::cuda::CudaKernel {
   int64_t max_ngram_size_;
   int64_t n_head_per_ngram_;
   T pad_id_;
-};
-
-template <typename T>
-class EngramGate final : public onnxruntime::cuda::CudaKernel {
- public:
-  explicit EngramGate(const OpKernelInfo& info);
-  Status ComputeInternal(OpKernelContext* context) const override;
-
- private:
-  float epsilon_;
 };
 
 }  // namespace cuda
