+++ onnxruntime/contrib_ops/cuda/bert/short_conv.h
@@ -1,0 +1,29 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "core/common/common.h"
+#include "core/providers/cuda/cuda_kernel.h"
+
+#include <string>
+
+namespace onnxruntime {
+namespace contrib {
+namespace cuda {
+
+template <typename T>
+class ShortConv final : public onnxruntime::cuda::CudaKernel {
+ public:
+  explicit ShortConv(const OpKernelInfo& info);
+  Status ComputeInternal(OpKernelContext* context) const override;
+
+ private:
+  std::string activation_;
+  int64_t dilation_;
+  float epsilon_;
+};
+
+}  // namespace cuda
+}  // namespace contrib
+}  // namespace onnxruntime
