+++ onnxruntime/contrib_ops/webgpu/bert/engram_gate.h
@@ -1,0 +1,42 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "core/providers/webgpu/program.h"
+#include "core/providers/webgpu/webgpu_kernel.h"
+
+namespace onnxruntime {
+namespace contrib {
+namespace webgpu {
+
+using onnxruntime::webgpu::ComputeContext;
+
+class EngramGateProgram final : public Program<EngramGateProgram> {
+ public:
+  EngramGateProgram(bool has_key_bias, bool has_value_bias)
+      : Program{"EngramGate"}, has_key_bias_(has_key_bias), has_value_bias_(has_value_bias) {}
+  Status GenerateShaderCode(ShaderHelper& shader) const override;
+  WEBGPU_PROGRAM_DEFINE_UNIFORM_VARIABLES({"total", ProgramUniformVariableDataType::Uint32},
+                                          {"hc_mult", ProgramUniformVariableDataType::Uint32},
+                                          {"hidden_size", ProgramUniformVariableDataType::Uint32},
+                                          {"embedding_size", ProgramUniformVariableDataType::Uint32},
+                                          {"epsilon", ProgramUniformVariableDataType::Float32});
+
+ private:
+  bool has_key_bias_;
+  bool has_value_bias_;
+};
+
+class EngramGate final : public WebGpuKernel {
+ public:
+  explicit EngramGate(const OpKernelInfo& info);
+  Status ComputeInternal(ComputeContext& context) const override;
+
+ private:
+  float epsilon_;
+};
+
+}  // namespace webgpu
+}  // namespace contrib
+}  // namespace onnxruntime
