+++ onnxruntime/contrib_ops/webgpu/bert/ngram_hash_mapping.h
@@ -1,0 +1,39 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "core/providers/webgpu/program.h"
+#include "core/providers/webgpu/webgpu_kernel.h"
+
+namespace onnxruntime {
+namespace contrib {
+namespace webgpu {
+
+using onnxruntime::webgpu::ComputeContext;
+
+class NgramHashMappingProgram final : public Program<NgramHashMappingProgram> {
+ public:
+  NgramHashMappingProgram() : Program{"NgramHashMapping"} {}
+  Status GenerateShaderCode(ShaderHelper& shader) const override;
+  WEBGPU_PROGRAM_DEFINE_UNIFORM_VARIABLES({"total", ProgramUniformVariableDataType::Uint32},
+                                          {"sequence_length", ProgramUniformVariableDataType::Uint32},
+                                          {"max_ngram_size", ProgramUniformVariableDataType::Uint32},
+                                          {"n_head_per_ngram", ProgramUniformVariableDataType::Uint32},
+                                          {"pad_id", ProgramUniformVariableDataType::Int32});
+};
+
+class NgramHashMapping final : public WebGpuKernel {
+ public:
+  explicit NgramHashMapping(const OpKernelInfo& info);
+  Status ComputeInternal(ComputeContext& context) const override;
+
+ private:
+  int64_t max_ngram_size_;
+  int64_t n_head_per_ngram_;
+  int64_t pad_id_;
+};
+
+}  // namespace webgpu
+}  // namespace contrib
+}  // namespace onnxruntime
